Created SpectralGraphTools
This file is the master file of the classes used to manipulate graphs and their spectra. It contains some plotting tools, as well as my own formulations of common and specialized spectral graph processing operations. At time of this commit, this file only contains one class, Graph_Spectrum, which plots the spectrum of a graph.